T-8: Reservation Form (replaces AddReservationModal)
- Add components/reservation-form.tsx — Dialog on desktop, Drawer on mobile
- Fields: guest name, party size, start/end time, table layout (TableBreakdownBuilder), notes
- table_breakdown stored as number[] and passed to createReservation/updateReservation
- Pre-populates tableBreakdown from editItem.table_breakdown in edit mode
- Delete components/add-reservation-modal.tsx; update CalendarDaySidebar and DayViewClient
